GRF-240: move test in back change autoload
